+++ src/main/java/com/maxmind/geoip2/model/IspResponse.java
@@ -3,6 +3,8 @@ package com.maxmind.geoip2.model;
 import com.fasterxml.jackson.annotation.JacksonInject;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.maxmind.db.MaxMindDbConstructor;
+import com.maxmind.db.MaxMindDbParameter;
 import com.maxmind.db.Network;
 import com.maxmind.geoip2.NetworkDeserializer;
 
@@ -39,6 +41,40 @@ public class IspResponse extends AsnResponse {
         super(autonomousSystemNumber, autonomousSystemOrganization, ipAddress, network);
         this.isp = isp;
         this.organization = organization;
+    }
+
+    @MaxMindDbConstructor
+    public IspResponse(
+            @MaxMindDbParameter(name="autonomous_system_number") Long autonomousSystemNumber,
+            @MaxMindDbParameter(name="autonomous_system_organization") String autonomousSystemOrganization,
+            @MaxMindDbParameter(name="ip_address") String ipAddress,
+            @MaxMindDbParameter(name="isp") String isp,
+            @MaxMindDbParameter(name="organization") String organization,
+            @MaxMindDbParameter(name="network") Network network
+    ) {
+        this(
+            autonomousSystemNumber != null ? autonomousSystemNumber.intValue() : null,
+            autonomousSystemOrganization,
+            ipAddress,
+            isp,
+            organization,
+            network
+        );
+    }
+
+    public IspResponse(
+            IspResponse response,
+            String ipAddress,
+            Network network
+    ) {
+        this(
+            response.getAutonomousSystemNumber(),
+            response.getAutonomousSystemOrganization(),
+            ipAddress,
+            response.getIsp(),
+            response.getOrganization(),
+            network
+        );
     }
 
     /**